Team-A-Project-02-project02-dev/src/main/java/org/revature/RevTaskManagement/service/ClientService.java
package org.revature.RevTaskManagement.service;

import org.revature.RevTaskManagement.models.Client;
import org.revature.RevTaskManagement.repository.ClientRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ClientService {

    @Autowired
    private ClientRepository clientRepository;

    public Client createClient(Client client) {
        return clientRepository.save(client);
    }

    public List<Client> getAllClients() {
        return clientRepository.findAll();
    }

}